refactor(openclaw): split plugin helper modules
